user creation
We can now create users with random id.

I have created a function that generates a random id by first generating the size of the id, we make sure the size is not < 10 with while loop. Then we use a for loop to generate a random number, looping through the size of the id.

results in a random length id than is not less than 10 numbers.