login function
Successfully implemented logging in.

After signing up we can attempt to login. For now we have a basic page saying that we are logged in.

Added authenticate function to register.js. This function is responsible for checking if the user exists with the use of cookies.
If user does not exist we redirect to sign up page.