major refactor
I have successfully and easily fixed the two pages.

Created two new classes, one for main content (.main-body) and another one for forms (.forms).

Main content has a margin on both sides of 20% and forms have margin on both sides of 40% and top margin of 10%. This puts the form in the middle of the page.