background and layout fixes
I have added a background color to the page to bring in more color as the page was fully white before.

Created a new color palette for the website.

Change button colors.

Fixed the way signup and login forms looked on mobile.

Had an issue where the background color would not cover the whole page. Fixed