footer complete
Basic footer has been created.

This footer contains a list of links to other pages on the website and a copyright.

Might change the color of the footer in the future as I feel that the footer sticks out too much and doe snot look too appealing.

Footer complete for the meantime.