delete notes
We can now successfully delete notes for the logged in user.

Refactoring of code needed to be done as I used jsoning library before which was a bad idea. It made everything more difficult and confusing.

Thanks to the use of the fs package we are nota able to read and write json files.