fully functioning bin
We can now restore notes. The deleted notes are moved to bin array in users profile, from there we render the binned notes. If the user presses delete on the note on bin page, the note will be permanently deleted, if the user presses the restore button, the note will be added back to notes and will be deleted form the bin array.