quick fix
Quickly refactored some code.

Could not figure out how to add a class to an item without the server crashing as js dom elements do not work in node and jquery does not want to work either.

Will have to look into this later.

Quickly just copied the sidemenu html code into start.hbs and bin.hbs. This is a temporary fix.