menu refactor + info
While adding more information I came across a small problem. The grid layout was not working properly.

I had to refactor some code for the grid to work once again.

This resulted in us having 3 notes in a row rather than 2 and having the grid cantered perfectly on the page without using margins.